+++ crates/adapters/src/lib.rs
@@ -147,6 +147,7 @@
 
 use bytemuck::NoUninit;
 use num_derive::FromPrimitive;
+use rustls::crypto::CryptoProvider;
 use serde::Serialize;
 
 mod catalog;
@@ -199,3 +200,15 @@ pub use transport::{
     AsyncErrorCallback, InputConsumer, InputEndpoint, InputReader, OutputEndpoint,
     TransportInputEndpoint,
 };
+
+/// Some dependencies of this crate use the `rustls` library. This library has two features
+/// `ring` and `aws-lc-rs`. When both are enabled, the library requires a process-wide default
+/// crypto provider to be configured. While no single dependency enables both these features,
+/// Rust's feature unification may end up enabling both of them, depending on the exact rustls
+/// dependency versions in use.
+///
+/// Bottom line: this function must be called in the `main` function in this crate, as well as
+/// in all tests that exercise libraries that use `rustls` internally.
+fn ensure_default_crypto_provider() {
+    let _ = CryptoProvider::install_default(rustls::crypto::aws_lc_rs::default_provider());
+}